# swagger/swagger.json
{"swagger":"2.0","info":{"version":""},"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/computing/add/{hash}/{private_key}":{"post":{"tags":["ComputingProvider"],"summary":"add ComputingProvider","description":"add computing resource","operationId":"ComputingProvider#add","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"hash","in":"path","description":"computing resource IPFS address","required":true,"type":"string"},{"name":"private_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}}}},"/computing/agree/{hash}/{ETH_key}/{request_id}":{"post":{"tags":["ComputingProvider"],"summary":"agree ComputingProvider","description":"agree computing request for request[ID]","operationId":"ComputingProvider#agree","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"hash","in":"path","required":true,"type":"string"},{"name":"request_id","in":"path","description":"request[ID]","required":true,"type":"integer"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}}}},"/computing/del/{hash}/{private_key}":{"post":{"tags":["ComputingProvider"],"summary":"del ComputingProvider","description":"delete computing resource","operationId":"ComputingProvider#del","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"hash","in":"path","description":"computing resource IPFS address","required":true,"type":"string"},{"name":"private_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}}}},"/computing/upload/{res_hash}/{aes_hash}/{ETH_key}/{request_id}":{"post":{"tags":["ComputingProvider"],"summary":"uploadRes ComputingProvider","description":"upload result hash for [request_id]","operationId":"ComputingProvider#uploadRes","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"aes_hash","in":"path","description":"encrypted aes key hash","required":true,"type":"string"},{"name":"request_id","in":"path","description":" [request_id]","required":true,"type":"integer"},{"name":"res_hash","in":"path","description":"encrypted result hash","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}}}},"/swagger-ui-dist/{filepath}":{"get":{"summary":"Download swagger-ui-dist/","operationId":"swagger-ui-dist#/swagger-ui-dist/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}}}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}}}}},"definitions":{"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}}}